update221104
- Trade 기능 구현 (미완)
- User Holding Stock 구현 (미완)

+++ src/main/java/com/stock/trade/StockTradeSessionDto.java
@@ -22,5 +22,5 @@ public class StockTradeSessionDto implements Serializable {
 	static final long serialVersionUID = 1L;
 	
 	private String stock_code;
-	private int stock_price;		
+	private double stock_price;		
 }
