+++ src/main/java/com/user/account/UserAccountDto.java
@@ -12,9 +12,9 @@ public class UserAccountDto {
 	// DB `account` table과 호환
 	private String serial_number; // 통장 번호
 	private String user_num;      // 사용자 번호
-	private int balance;          // 원화
-	private int us;               // 달러
-	private int euro;             // 유로
-	private int jpy;              // 엔화
-	private int yuan;             // 위안
+	private double balance;          // 원화
+	private double us;               // 달러
+	private double euro;             // 유로
+	private double jpy;              // 엔화
+	private double yuan;             // 위안
 }
